Update 01.md
I read the first chapter and encountered a few typos. I also took the liberty of changing things here and there - where I found the change had the effect on the reading of the sentence - making it smoother for me. Corrections can be used outside of the PR flow or course;). Glad to help, good read this one!